Resolved all the TS error and also restructured the folder structure completely
