--- .env
@@ -1,9 +1,0 @@
-VITE_API_KEY=AIzaSyBdM_imftB9Qxcgb67ZEfIgnsmxusZVELo
-VITE_AUTH_DOMAIN=satkriya.firebaseapp.com
-VITE_PROJECT_ID=satkriya
- VITE_STORAGE_BUCKET=satkriya.firebasestorage.app
-VITE_MESSAGING_SENDER_ID=36171792594
-VITE_APP_ID=1:36171792594:web:23e21cab6d170ff713ad0b
-
-VITE_CLOUDINARY_PRESET=fu6t0ggw
-VITE_CLOUDINARY_NAME=daawurvug
