+++ src/components/AuthLayoutFarmer/AuthLayoutFarmer.types.ts
@@ -1,0 +1,9 @@
+import { ReactNode } from "react";
+
+interface AuthLayoutFarmerProps {
+  children: ReactNode;
+  title: string;
+  subtitle: string;
+}
+
+export default AuthLayoutFarmerProps;
