+++ src/components/BreedCard/BreedCard.types.ts
@@ -1,0 +1,8 @@
+interface BreedCardProps {
+  name: string;
+  image: string;
+  description: string;
+  traits: string[];
+}
+
+export default BreedCardProps;
