--- src/components/Forms/Farmer/FarmerLoginForm.tsx
@@ -1,6 +1,0 @@
-
-const FarmerLoginForm = () => {
-  return <div>FarmerLoginForm</div>;
-};
-
-export default FarmerLoginForm;
