+++ src/components/Input/Input.tsx
@@ -1,11 +1,10 @@
-import React from "react";
+import { FC } from "react";
+import InputProps from "./Input.types";
 
-export function Input({ label, error, className = "", ...props }) {
+const Input: FC<InputProps> = ({ label, error, className = "", ...props }) => {
   return (
     <div className="space-y-1">
-      <label className="block text-sm font-medium text-gray-700">
-        {label}
-      </label>
+      <label className="block text-sm font-medium text-gray-700">{label}</label>
       <input
         className={`
           w-full px-3 py-2 border rounded-lg shadow-sm
@@ -18,4 +17,6 @@ export function Input({ label, error, className = "", ...props }) {
       {error && <p className="text-sm text-red-600">{error}</p>}
     </div>
   );
-}
+};
+
+export default Input;
