+++ src/components/Input/Input.types.ts
@@ -1,0 +1,8 @@
+interface InputProps {
+  label: string;
+  error: string;
+  className: string;
+  [key: string]: any;
+}
+
+export default InputProps;
