+++ src/constants/breeds.ts
@@ -1,0 +1,25 @@
+const breeds = [
+  {
+    name: "Gir",
+    image:
+      "https://images.unsplash.com/photo-1570042225831-d98fa7577f1e?ixlib=rb-4.0.3",
+    description: "Known for high milk yield and disease resistance",
+    traits: ["A2 Milk", "Heat Tolerant", "Disease Resistant"],
+  },
+  {
+    name: "Sahiwal",
+    image:
+      "https://images.unsplash.com/photo-1545468800-85cc9bc6ecf7?ixlib=rb-4.0.3",
+    description: "Excellent dairy breed with good mothering ability",
+    traits: ["High Fat Content", "Adaptable", "Good Mother"],
+  },
+  {
+    name: "Red Sindhi",
+    image:
+      "https://encrypted-tbn2.gstatic.com/images?q=tbn:ANd9GcQ7QVzl5Goiu3fUqO6vh7Xs1H59QlBphyYIRN_J2ovWaaYKb8LPLm5menceV-gl_3gioDKE3bFiy8teUlqtGk6vHg",
+    description: "Hardy breed suitable for tropical climate",
+    traits: ["Heat Resistant", "Low Maintenance", "Regular Breeder"],
+  },
+];
+
+export default breeds;
