+++ src/constants/impactMetrics.ts
@@ -1,0 +1,9 @@
+const impactMetrics = [
+  { label: "Cows Saved", value: "12,450+" },
+  { label: "Farmers Supported", value: "5,200+" },
+  { label: "Successful Treatments", value: "28,900+" },
+  { label: "Organic Farms", value: "850+" },
+  { label: "CO₂ Reduced (tons)", value: "15,000+" },
+];
+
+export default impactMetrics;
