+++ src/pages/Expert/LoginExpert/LoginExpert.tsx
@@ -1,14 +1,14 @@
-import { AuthLayoutFarmer } from "@/components/AuthLayoutFarmer";
+import AuthLayoutExpert from "@/components/AuthLayoutExpert/AuthLayoutExpert";
 import ExpertLoginForm from "@/components/Forms/Expert/ExpertLoginForm/ExpertLoginForm";
 
 const LoginExpert = () => {
   return (
-    <AuthLayoutFarmer
+    <AuthLayoutExpert
       title="Welcome Back!"
       subtitle="Sign in to your account to continue your journey with us."
     >
       <ExpertLoginForm />
-    </AuthLayoutFarmer>
+    </AuthLayoutExpert>
   );
 };
